--- config/go.mod
@@ -1,5 +1,0 @@
-module config
-
-go 1.21.1
-
-require github.com/joho/godotenv v1.5.1
