--- firebase/go.mod
@@ -1,17 +1,0 @@
-module github.com/ReticentFacade/termvoid/firebase
-
-go 1.21.1
-
-// require github.com/joho/godotenv v1.5.1
-
-require (
-	github.com/ReticentFacade/termvoid/pkg v0.0.0-20230925041327-3d36a431a642 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	// github.com/joho/godotenv v1.5.1 // indirect
-	// golang.org/x/net v0.15.0 // indirect
-	golang.org/x/sys v0.12.0 // indirect
-	golang.org/x/text v0.13.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20231002182017-d307bd883b97 // indirect
-	google.golang.org/grpc v1.58.2 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
