--- go.work
@@ -1,8 +1,0 @@
-go 1.21.1
-
-use (
-	./firebase
-	./pkg/proto
-	./void
-	./config
-)
