+++ main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"termvoid/cmd"
+	"termvoid/server"
+)
+
+func main() {
+	cmd.Execute()
+	server.ServerStartup()
+}
