--- pkg/go.mod
@@ -1,3 +1,0 @@
-module pkg
-
-go 1.21.1
