+++ utils/directory.go
@@ -1,0 +1,16 @@
+package utils
+
+import (
+	"os"
+) 
+
+func IfDirExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil // Exists
+	}
+	if os.IsNotExist(err) {
+		return false, nil // Doesn't exist
+	}
+	return false, err // Error while checking
+}
