--- void/cgo_integration1.txt
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"net/http"
-	"void/cmd"
-	"void/localhost"
-)
-
-func cgo_integration_test1() {
-	// CLI
-	cmd.Execute()
-
-	fmt.Println("main.go: Starting deleteit...")
-	deleteit()
-
-	// ===============================
-	// Localhost server
-	fileServer := http.FileServer(http.Dir("./localhost"))
-	http.Handle("/", fileServer)
-	http.HandleFunc("/upload", localhost.UploadingFunc)
-
-	// Start the server
-	port := 8080
-	fmt.Println("main.go: Server started at port:", port)
-
-	err := http.ListenAndServe(fmt.Sprintf(":%d", port), nil)
-	if err != nil {
-		fmt.Println("Error while starting server:", err)
-		return
-	}
-}
