--- void/go.mod
@@ -1,22 +1,0 @@
-module void
-
-go 1.21.1
-
-require (
-	github.com/ReticentFacade/termvoid/pkg v0.0.0-20230925041327-3d36a431a642
-	github.com/google/uuid v1.3.1
-	// github.com/joho/godotenv v1.5.1
-	github.com/spf13/cobra v1.7.0
-	golang.org/x/net v0.15.0
-	google.golang.org/grpc v1.58.2
-)
-
-require (
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.12.0 // indirect
-	golang.org/x/text v0.13.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20230711160842-782d3b101e98 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
