--- void/main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"void/cmd"
-	"void/server"
-)
-
-func main() {
-	cmd.Execute()
-	server.ServerStartup()
-}
